Refactor: Network module configuration
- Renamed `networkRetrofitModule` to `networkModule`.
- Moved Retrofit client creation logic to `createRetrofitClient.kt`.
- Updated module to use constants for base URL and access token.
- Standardized qualifiers for Retrofit builders.
- Simplified Retrofit service provisioning.
