+++ app/src/main/java/com/giraffe/cineverseapp/CineVerseApp.kt
@@ -8,7 +8,7 @@ import androidx.work.WorkManager
 import com.giraffe.cineverseapp.di.databaseModule
 import com.giraffe.cineverseapp.di.featureApiModule
 import com.giraffe.cineverseapp.di.localDataSourceModule
-import com.giraffe.cineverseapp.di.networkRetrofitModule
+import com.giraffe.cineverseapp.di.networkModule
 import com.giraffe.cineverseapp.di.repositoryModule
 import com.giraffe.cineverseapp.di.useCaseModule
 import com.giraffe.cineverseapp.di.viewModelModule
@@ -42,7 +42,7 @@ class CineVerseApp : Application() {
                 viewModelModule,
                 imageViewerModule,
                 featureApiModule,
-                networkRetrofitModule
+                networkModule
             )
         }
         setupCacheCleanupWorker()
