+++ app/src/main/java/com/giraffe/cineverseapp/di/networkModule.kt
@@ -1,6 +1,9 @@
 package com.giraffe.cineverseapp.di
 
-import com.giraffe.cineverseapp.BuildConfig
+import com.giraffe.cineverseapp.BuildConfig.ACCESS_TOKEN
+import com.giraffe.cineverseapp.BuildConfig.BASE_URL
+import com.giraffe.cineverseapp.data.network.createRetrofitClient
+import com.giraffe.cineverseapp.data.network.provideRetrofitService
 import com.giraffe.media.explore.datasource.remote.ExploreRemoteDataSource
 import com.giraffe.media.explore.retrofit.ExploreApiServiceRetrofit
 import com.giraffe.media.explore.retrofit.ExploreRemoteDataSourceImplRetrofit
@@ -13,7 +16,6 @@ import com.giraffe.media.person.retrofit.PersonRemoteDataSourceImplRetrofit
 import com.giraffe.media.series.datasource.remote.SeriesRemoteDataSource
 import com.giraffe.media.series.retrofit.SeriesApiServiceRetrofit
 import com.giraffe.media.series.retrofit.SeriesRemoteRetrofitDataSourceImp
-import com.giraffe.media.util.AuthInterceptor
 import com.giraffe.media.util.RetrofitRequestBuilder
 import com.giraffe.repository.SessionManagerImpl
 import com.giraffe.repository.datasource.UserRemoteDataSource
@@ -23,21 +25,21 @@ import com.giraffe.user.retrofit.UserRemoteDataSourceImplRetrofit
 import com.jakewharton.retrofit2.converter.kotlinx.serialization.asConverterFactory
 import kotlinx.serialization.json.Json
 import okhttp3.MediaType.Companion.toMediaType
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
 import org.koin.core.module.dsl.singleOf
 import org.koin.core.qualifier.named
 import org.koin.dsl.bind
 import org.koin.dsl.module
 import retrofit2.Retrofit
 
-private const val BASE_URL = "BASE_URL"
-private const val ACCESS_TOKEN = "ACCESS_TOKEN"
+private const val QUALIFIER_MOVIES_BUILDER = "movies_builder"
+private const val QUALIFIER_SERIES_BUILDER = "series_builder"
+private const val QUALIFIER_EXPLORE_BUILDER = "explore_builder"
+private const val QUALIFIER_PERSON_BUILDER = "person_builder"
 
-val networkRetrofitModule = module {
+val networkModule = module {
 
-    single(named(BASE_URL)) { BuildConfig.BASE_URL }
-    single(named(ACCESS_TOKEN)) { BuildConfig.ACCESS_TOKEN }
+    single(named(BASE_URL)) { BASE_URL }
+    single(named(ACCESS_TOKEN)) { ACCESS_TOKEN }
 
     single {
         createRetrofitClient(get(named(ACCESS_TOKEN)))
@@ -61,46 +63,40 @@ val networkRetrofitModule = module {
     single { provideRetrofitService<ExploreApiServiceRetrofit>(get()) }
     single { provideRetrofitService<PersonApiServiceRetrofit>(get()) }
 
-    single<RetrofitRequestBuilder<MoviesApiServiceRetrofit>>(named("movies_builder")) {
+    single<RetrofitRequestBuilder<MoviesApiServiceRetrofit>>(named(QUALIFIER_MOVIES_BUILDER)) {
         val api = get<Retrofit>().create(MoviesApiServiceRetrofit::class.java)
         RetrofitRequestBuilder(api)
     }
 
-    single<RetrofitRequestBuilder<SeriesApiServiceRetrofit>>(named("series_builder")) {
+    single<RetrofitRequestBuilder<SeriesApiServiceRetrofit>>(named(QUALIFIER_SERIES_BUILDER)) {
         val api = get<Retrofit>().create(SeriesApiServiceRetrofit::class.java)
         RetrofitRequestBuilder(api)
     }
 
-    single<RetrofitRequestBuilder<ExploreApiServiceRetrofit>>(named("explore_builder")) {
+    single<RetrofitRequestBuilder<ExploreApiServiceRetrofit>>(named(QUALIFIER_EXPLORE_BUILDER)) {
         val api = get<Retrofit>().create(ExploreApiServiceRetrofit::class.java)
         RetrofitRequestBuilder(api)
     }
-    single<RetrofitRequestBuilder<UserApiServiceRetrofit>>(named("user_builder")) {
-        val api = get<Retrofit>().create(UserApiServiceRetrofit::class.java)
-        RetrofitRequestBuilder(api)
-    }
-
-
     single<ExploreRemoteDataSource> {
-        ExploreRemoteDataSourceImplRetrofit(get(named("explore_builder")))
+        ExploreRemoteDataSourceImplRetrofit(get(named(QUALIFIER_EXPLORE_BUILDER)))
     }
-    single<RetrofitRequestBuilder<PersonApiServiceRetrofit>>(named("person_builder")) {
+    single<RetrofitRequestBuilder<PersonApiServiceRetrofit>>(named(QUALIFIER_PERSON_BUILDER)) {
         val api = get<Retrofit>().create(PersonApiServiceRetrofit::class.java)
         RetrofitRequestBuilder(api)
     }
     single<MoviesRemoteDataSource> {
-        MoviesRemoteDataSourceImplRetrofit(get(named("movies_builder")))
+        MoviesRemoteDataSourceImplRetrofit(get(named(QUALIFIER_MOVIES_BUILDER)))
     }
     single<PersonRemoteDataSource> {
-        PersonRemoteDataSourceImplRetrofit(get(named("person_builder")))
+        PersonRemoteDataSourceImplRetrofit(get(named(QUALIFIER_PERSON_BUILDER)))
     }
     single<SeriesRemoteDataSource> {
-        SeriesRemoteRetrofitDataSourceImp(get(named("series_builder")))
+        SeriesRemoteRetrofitDataSourceImp(get(named(QUALIFIER_SERIES_BUILDER)))
     }
 
     single(named("user_retrofit")) {
         Retrofit.Builder()
-            .baseUrl("https://api.themoviedb.org/3/")
+            .baseUrl(BASE_URL)
             .addConverterFactory(get<Json>().asConverterFactory("application/json".toMediaType()))
             .client(get())
             .build()
@@ -117,17 +113,6 @@ val networkRetrofitModule = module {
 
 }
 
-private fun createRetrofitClient(accessToken: String): OkHttpClient {
-    val loggingInterceptor = HttpLoggingInterceptor().apply {
-        level = if (BuildConfig.DEBUG) HttpLoggingInterceptor.Level.BODY else HttpLoggingInterceptor.Level.NONE
-    }
 
-    return OkHttpClient.Builder()
-        .addInterceptor(AuthInterceptor(accessToken))
-        .addInterceptor(loggingInterceptor)
-        .build()
-}
 
-inline fun <reified T> provideRetrofitService(retrofit: Retrofit): T {
-    return retrofit.create(T::class.java)
-}
+
